# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Saude+API",
    "version": "0.1.0"
  },
  "paths": {
    "/eventos/": {
      "get": {
        "summary": "Listar Eventos",
        "operationId": "listar_eventos_eventos__get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Eventos_list_Schema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/eventos/{id_evento}": {
      "get": {
        "summary": "Listar Eventos Por Id",
        "operationId": "listar_eventos_por_id_eventos__id_evento__get",
        "parameters": [
          {
            "name": "id_evento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Evento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evento_retorno_Schema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/eventos/criar_evento/": {
      "post": {
        "summary": "Criar Eventos",
        "operationId": "criar_eventos_eventos_criar_evento__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Evento_Schema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evento_retorno_Schema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/eventos/atualizar_evento/{id_evento}": {
      "put": {
        "summary": "Atualizar Evento",
        "operationId": "atualizar_evento_eventos_atualizar_evento__id_evento__put",
        "parameters": [
          {
            "name": "id_evento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Evento"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Evento_Schema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evento_retorno_Schema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/eventos/remover_evento/{id_evento}": {
      "delete": {
        "summary": "Remover Evento",
        "operationId": "remover_evento_eventos_remover_evento__id_evento__delete",
        "parameters": [
          {
            "name": "id_evento",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Evento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/comum/": {
      "get": {
        "summary": "Listar Usuarios Comum",
        "operationId": "listar_usuarios_comum_usuarios_comum__get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario_comum_lista_Schema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/comum/{id_comum}": {
      "get": {
        "summary": "Listar Usuario Comum Por Id",
        "operationId": "listar_usuario_comum_por_id_usuarios_comum__id_comum__get",
        "parameters": [
          {
            "name": "id_comum",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Comum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario_comum_retorno_Schema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/comum/criar_evento/": {
      "post": {
        "summary": "Criar Usuario Comum",
        "operationId": "criar_usuario_comum_usuarios_comum_criar_evento__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario_comum"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario_comum_retorno_Schema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/comum/atualizar_usuario/{id_usuario}": {
      "put": {
        "summary": "Atualizar Usuario Comum",
        "operationId": "atualizar_usuario_comum_usuarios_comum_atualizar_usuario__id_usuario__put",
        "parameters": [
          {
            "name": "id_usuario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Usuario"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario_comum"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario_comum_retorno_Schema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/comum/deletar_usuario/{id_usuario}": {
      "delete": {
        "summary": "Remover Usuario Comum",
        "operationId": "remover_usuario_comum_usuarios_comum_deletar_usuario__id_usuario__delete",
        "parameters": [
          {
            "name": "id_usuario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Usuario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Comentario_retorno_Schema": {
        "properties": {
          "nota": {
            "type": "integer",
            "title": "Nota"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "conteudo": {
            "type": "string",
            "title": "Conteudo"
          },
          "usuario": {
            "type": "integer",
            "title": "Usuario"
          },
          "evento": {
            "type": "integer",
            "title": "Evento"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "data_hora_feito": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora Feito"
          }
        },
        "type": "object",
        "required": [
          "nota",
          "titulo",
          "conteudo",
          "usuario",
          "evento",
          "id",
          "data_hora_feito"
        ],
        "title": "Comentario_retorno_Schema"
      },
      "Evento_Schema": {
        "properties": {
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "descricao": {
            "type": "string",
            "title": "Descricao"
          },
          "data_hora_evento": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora Evento"
          },
          "data_hora_fim": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora Fim"
          },
          "publico_alvo": {
            "type": "string",
            "title": "Publico Alvo"
          },
          "categoria": {
            "type": "integer",
            "title": "Categoria"
          },
          "foto_evento": {
            "type": "integer",
            "title": "Foto Evento"
          },
          "criador_institucional": {
            "type": "integer",
            "title": "Criador Institucional"
          },
          "capacidade_maxima": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacidade Maxima"
          },
          "unidade_associada": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unidade Associada"
          },
          "endereco": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endereco"
          },
          "pagina_evento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pagina Evento"
          }
        },
        "type": "object",
        "required": [
          "titulo",
          "descricao",
          "data_hora_evento",
          "data_hora_fim",
          "publico_alvo",
          "categoria",
          "foto_evento",
          "criador_institucional"
        ],
        "title": "Evento_Schema"
      },
      "Evento_retorno_Schema": {
        "properties": {
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "descricao": {
            "type": "string",
            "title": "Descricao"
          },
          "data_hora_evento": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora Evento"
          },
          "data_hora_fim": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora Fim"
          },
          "publico_alvo": {
            "type": "string",
            "title": "Publico Alvo"
          },
          "categoria": {
            "type": "integer",
            "title": "Categoria"
          },
          "foto_evento": {
            "type": "integer",
            "title": "Foto Evento"
          },
          "criador_institucional": {
            "type": "integer",
            "title": "Criador Institucional"
          },
          "capacidade_maxima": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacidade Maxima"
          },
          "unidade_associada": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unidade Associada"
          },
          "endereco": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endereco"
          },
          "pagina_evento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pagina Evento"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "inscricoes_atuais": {
            "type": "integer",
            "title": "Inscricoes Atuais",
            "default": 0
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "data_hora_criacao": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora Criacao"
          },
          "data_hora_cancelamento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Hora Cancelamento"
          },
          "data_hora_ultima_atualizacao": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Hora Ultima Atualizacao"
          },
          "comentarios": {
            "items": {
              "$ref": "#/components/schemas/Comentario_retorno_Schema"
            },
            "type": "array",
            "title": "Comentarios",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "titulo",
          "descricao",
          "data_hora_evento",
          "data_hora_fim",
          "publico_alvo",
          "categoria",
          "foto_evento",
          "criador_institucional",
          "id",
          "status",
          "data_hora_criacao"
        ],
        "title": "Evento_retorno_Schema"
      },
      "Eventos_list_Schema": {
        "properties": {
          "eventos": {
            "items": {
              "$ref": "#/components/schemas/Evento_retorno_Schema"
            },
            "type": "array",
            "title": "Eventos"
          }
        },
        "type": "object",
        "required": [
          "eventos"
        ],
        "title": "Eventos_list_Schema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MessageSchema": {
        "properties": {
          "mensagem": {
            "type": "string",
            "title": "Mensagem"
          }
        },
        "type": "object",
        "required": [
          "mensagem"
        ],
        "title": "MessageSchema"
      },
      "Usuario_comum": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "cpf": {
            "type": "string",
            "title": "Cpf"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "senha": {
            "type": "string",
            "title": "Senha"
          },
          "data_nascimento": {
            "type": "string",
            "format": "date-time",
            "title": "Data Nascimento"
          },
          "telefone": {
            "type": "string",
            "title": "Telefone"
          },
          "foto_perfil": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foto Perfil"
          },
          "regiao_preferida": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regiao Preferida"
          }
        },
        "type": "object",
        "required": [
          "email",
          "cpf",
          "nome",
          "senha",
          "data_nascimento"
        ],
        "title": "Usuario_comum"
      },
      "Usuario_comum_lista_Schema": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "senha": {
            "type": "string",
            "title": "Senha"
          },
          "telefone": {
            "type": "string",
            "title": "Telefone"
          },
          "cpf": {
            "type": "string",
            "title": "Cpf"
          },
          "data_nascimento": {
            "type": "string",
            "format": "date-time",
            "title": "Data Nascimento"
          },
          "regiao_preferida": {
            "type": "string",
            "title": "Regiao Preferida"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "usuarios": {
            "items": {
              "$ref": "#/components/schemas/Usuario_comum_retorno_Schema"
            },
            "type": "array",
            "title": "Usuarios"
          }
        },
        "type": "object",
        "required": [
          "email",
          "nome",
          "senha",
          "telefone",
          "cpf",
          "data_nascimento",
          "regiao_preferida",
          "id",
          "usuarios"
        ],
        "title": "Usuario_comum_lista_Schema"
      },
      "Usuario_comum_retorno_Schema": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "senha": {
            "type": "string",
            "title": "Senha"
          },
          "telefone": {
            "type": "string",
            "title": "Telefone"
          },
          "cpf": {
            "type": "string",
            "title": "Cpf"
          },
          "data_nascimento": {
            "type": "string",
            "format": "date-time",
            "title": "Data Nascimento"
          },
          "regiao_preferida": {
            "type": "string",
            "title": "Regiao Preferida"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "email",
          "nome",
          "senha",
          "telefone",
          "cpf",
          "data_nascimento",
          "regiao_preferida",
          "id"
        ],
        "title": "Usuario_comum_retorno_Schema"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}